iot/battle-mlx-cam/back/requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
Pillow>=10.0.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0

rembg>=2.0.50
torch>=2.0.0
torchvision>=0.15.0
flask>=3.0.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
pytest>=7.0.0
